requirements.txt
prettytable==0.*
requests
yattag==1.*
pyserial>2.5
jsonmerge>=1.4.0,==1.*
jsonschema<3.0.0
mbed-ls>=1.5.1,==1.*
semver==2.*
mbed-flasher>=0.10.1,==0.10.*
six==1.*
pydash==4.*
transitions==0.*